move output code out of the tree
