+++ sway/meson.build
@@ -86,7 +86,6 @@ sway_sources = files(
 	'security.c',
 	'tree/container.c',
 	'tree/layout.c',
-	'tree/output.c',
 	'tree/view.c',
 	'tree/workspace.c',
 )
